reafactoring round one
