+++ packages/voucherify/data/get-redeemmables-for-validation.ts
@@ -1,0 +1,5 @@
+export const getRedeemmablesForValidation = (couponCodes: string[]) =>
+  couponCodes.map((couponCode) => ({
+    id: couponCode,
+    object: 'voucher' as const,
+  }))
