+++ packages/voucherify/data/has-at-least-one-redeemable.ts
@@ -1,0 +1,6 @@
+import { getCartDiscounts } from './persit'
+
+export const hasAtLeastOneRedeemable = async (cartId: string) => {
+  const cartDiscountsStorage = await getCartDiscounts(cartId)
+  return cartDiscountsStorage && cartDiscountsStorage.length > 0
+}
