+++ packages/voucherify/package.json
@@ -10,7 +10,9 @@
     "ts": "tsc --noEmit --incremental"
   },
   "dependencies": {
-    "@composable/types": "workspace:*"
+    "@composable/types": "workspace:*",
+    "@types/node-persist": "^3.1.5",
+    "node-persist": "^3.1.3"
   },
   "devDependencies": {
     "eslint-config-custom": "workspace:*",
